pyproject.toml
[tool.poetry]
name = "motionpicture"
version = "0.1.2"
description = "Make animations with Python"
authors = ["Gabriele Bozzola <gabrielebozzola@arizona.edu>"]
maintainers = ["Gabriele Bozzola <gabrielebozzola@arizona.edu>"]
license = "GPL-3.0-or-later"
readme = "README.md"
homepage = "https://github.com/sbozzolo/motionpicture"
repository = "https://github.com/sbozzolo/motionpicture"
documentation = "https://sbozzolo.github.io/motionpicture"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/sbozzolo/motionpicture/issues"

[tool.poetry.dependencies]
python = "^3.6"
ffmpeg-python = "^0.1.0"
ConfigArgParse = "^1.0.0"
tqdm = "^4.0.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pytest = "^6.2.1"
pytest-cov = "^2.10.1"
matplotlib = "^3.3.3"

[tool.poetry.scripts]
mopi = "motionpicture.mopi:main"

[tool.black]
line-length = 89
exclude = '''
/(
    \.git
  | docs
)/
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
